pyproject.toml
[project]
name = "se-toolkit-hackathon"
version = "0.1.0"
description = "Psychological testing Telegram bot"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=20.0",
]
